+++ src/Translator/TranslationsLoaderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yiisoft\I18n;
+namespace Yiisoft\I18n\Translator;
 
 interface TranslationsLoaderInterface
 {
