+++ tests/TranslatorTest.php
@@ -5,7 +5,7 @@ namespace Yii\I18n\Tests;
 use PHPUnit\Framework\TestCase;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Yiisoft\I18n\Event\MissingTranslationEvent;
-use Yiisoft\I18n\TranslationsLoaderInterface;
+use Yiisoft\I18n\Translator\TranslationsLoaderInterface;
 use Yiisoft\I18n\Translator\Translator;
 use Yiisoft\I18n\TranslatorInterface;
